engine: Upper bound for VmPool.MaxAssignedVmsPerUser
Added an upper bound to the VmPool.MaxAssignedVmsPerUser property of
32,767 to prevent integer overflows and the possibility to get VmPool
in incosistent state by submitting exceedingly large value reulting in
negative value being set.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=988056

+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/models/vms/PoolModelBehaviorBase.java
@@ -304,7 +304,7 @@ public abstract class PoolModelBehaviorBase extends VmModelBehaviorBase<PoolMode
                 new IValidation[]
                 {
                         new NotEmptyValidation(),
-                        new IntegerValidation(1, Integer.MAX_VALUE)
+                        new IntegerValidation(1, Short.MAX_VALUE)
                 });
 
         getModel().setIsGeneralTabValid(getModel().getIsGeneralTabValid()
